+++ dymaptic.Pro.ZoomToCoordinates/Config.daml
@@ -37,7 +37,7 @@
           <tooltip heading="Get Coordinates">
 				  Click the map to get coordinates in your chosen coordinate format.<disabledText /></tooltip>
         </tool>
-        <button id="dymaptic_Pro_ZoomToCoordinates_Views_SettingsView" caption="Settings" className="dymaptic.Pro.ZoomToCoordinates.Views.ShowSettingsView" loadOnClick="true" smallImage="CogWheel16" largeImage="CogWheel32">
+        <button id="dymaptic_Pro_ZoomToCoordinates_Views_SettingsView" caption="Settings" className="dymaptic.Pro.ZoomToCoordinates.Views.ShowSettingsButton" loadOnClick="true" smallImage="CogWheel16" largeImage="CogWheel32">
           <tooltip heading="Setting">Customize startup and graphic settings if graphic requested.<disabledText /></tooltip>
         </button>
         <button id="dymaptic_Pro_ZoomToCoordinates_Views_About" caption="About" className="dymaptic.Pro.ZoomToCoordinates.Views.ShowAbout" loadOnClick="true" smallImage="Images/dymaptic-logo---circle.png" largeImage="Images/dymaptic-logo---circle.png">
