+++ dymaptic.Pro.ZoomToCoordinates/LatitudeBandHelper.cs
@@ -3,45 +3,33 @@
 namespace dymaptic.Pro.ZoomToCoordinates;
 public class LatitudeBandHelper
 {
-    // Approximate number of meters per latitude band (8 degrees latitude)
-    private const int BandHeightMeters = 888000;
-
-    // Latitude bands (C to X, excluding I and O)
-    private static readonly string LatitudeBands = "CDEFGHJKLMNPQRSTUVWX";
-    private static char SouthernmostBand => 'C';
-    private static char NorthernmostBand => 'X';
-
-    // Returns true if band is in the Northern Hemisphere
-    private static bool IsNorthernHemisphere(char band)
-    {
-        return band >= 'N'; // 'N' is the first band above the Equator
-    }
-
     public static Dictionary<char, int> GetLatitudeBandStartNorthings()
     {
-        var bandNorthings = new Dictionary<char, int>();
-
-        for (int i = 0; i < LatitudeBands.Length; i++)
+        if (_bandNorthings is null)
         {
-            char band = LatitudeBands[i];
+            _bandNorthings = new Dictionary<char, int>();
+            for (int i = 0; i < LatitudeBands.Length; i++)
+            {
+                char band = LatitudeBands[i];
 
-            if (IsNorthernHemisphere(band))
-            {
-                // Northern Hemisphere: start at 0m (Equator), increase northward
-                int northBandIndex = i - LatitudeBands.IndexOf('N');
-                int northing = northBandIndex * BandHeightMeters;
-                bandNorthings[band] = northing;
-            }
-            else
-            {
-                // Southern Hemisphere: start at 10,000,000m (Equator), decrease southward
-                int southBandIndex = LatitudeBands.IndexOf('M') - i + 1;
-                int northing = 10000000 - (southBandIndex * BandHeightMeters);
-                bandNorthings[band] = northing;
+                if (IsNorthernHemisphere(band))
+                {
+                    // Northern Hemisphere: start at 0m (Equator), increase northward
+                    int northBandIndex = i - LatitudeBands.IndexOf('N');
+                    int northing = northBandIndex * BandHeightMeters;
+                    _bandNorthings[band] = northing;
+                }
+                else
+                {
+                    // Southern Hemisphere: start at 10,000,000m (Equator), decrease southward
+                    int southBandIndex = LatitudeBands.IndexOf('M') - i + 1;
+                    int northing = 10000000 - (southBandIndex * BandHeightMeters);
+                    _bandNorthings[band] = northing;
+                }
             }
         }
 
-        return bandNorthings;
+        return _bandNorthings;
     }
 
     /// <summary>
@@ -66,4 +54,19 @@ public class LatitudeBandHelper
 
         return currentNorthing + offset;
     }
+
+    // Returns true if band is in the Northern Hemisphere
+    private static bool IsNorthernHemisphere(char band)
+    {
+        return band >= 'N'; // 'N' is the first band above the Equator
+    }
+
+    // Approximate number of meters per latitude band (8 degrees latitude)
+    private const int BandHeightMeters = 888000;
+
+    // Latitude bands (C to X, excluding I and O)
+    private static readonly string LatitudeBands = "CDEFGHJKLMNPQRSTUVWX";
+    private static char SouthernmostBand => 'C';
+    private static char NorthernmostBand => 'X';
+    private static Dictionary<char, int>? _bandNorthings;
 }
