+++ dymaptic.Pro.ZoomToCoordinates/ViewModels/SettingsViewModel.cs
@@ -1,6 +1,5 @@
 ﻿using ArcGIS.Desktop.Framework.Contracts;
 using System;
-using System.Collections.ObjectModel;
 
 namespace dymaptic.Pro.ZoomToCoordinates.ViewModels;
 
@@ -11,13 +10,13 @@ internal class SettingsViewModel : Page
 		ZoomToCoordinatesModule.Current.SettingsLoaded += Current_SettingsLoaded;
 	}
 
-	public ObservableCollection<string> MarkerSchemes { get; set; } = ["Circle", "Cross", "Diamond", "Square", "X", "Triangle", "Pushpin", "Star", "RoundedSquare", "RoundedTriangle", "Rod", "Rectangle", "RoundedRectangle", "Hexagon", "StretchedHexagon", "RaceTrack", "HalfCircle", "Cloud"];
+	public string[] MarkerSchemes { get; set; } = ["Circle", "Cross", "Diamond", "Square", "X", "Triangle", "Pushpin", "Star", "RoundedSquare", "RoundedTriangle", "Rod", "Rectangle", "RoundedRectangle", "Hexagon", "StretchedHexagon", "RaceTrack", "HalfCircle", "Cloud"];
 
-	public ObservableCollection<string> ColorSchemes { get; set; } = ["Black", "Gray", "White", "Red", "Green", "Blue"];
+	public string[] ColorSchemes { get; set; } = ["Black", "Gray", "White", "Red", "Green", "Blue"];
 
-	public ObservableCollection<string> FontFamilySchemes { get; set; } = ["Arial", "Broadway", "Papyrus", "Tahoma", "Times New Roman"];
+	public string[] FontFamilySchemes { get; set; } = ["Arial", "Broadway", "Papyrus", "Tahoma", "Times New Roman"];
 
-	public ObservableCollection<string> FontStyleSchemes { get; set; } = ["Regular", "Bold", "Italic"];
+	public string[] FontStyleSchemes { get; set; } = ["Regular", "Bold", "Italic"];
 
 	public double Longitude
 	{
